refactor: separate `SubscriberCreate` page into modules
