+++ frontend/src/lib/utils.ts
@@ -1,0 +1,3 @@
+export function toHex(v: number | undefined): string {
+  return ("00" + v?.toString(16).toUpperCase()).substr(-2);
+}
